Fix potential 32-bit overflow in cell count methods
Cast to uint64_t before multiplying in cellRowCount() and
cellColumnCount() to prevent overflow at high levels with large bounds.

+++ marine_autonomy/include/marine_autonomy/gggs/bounds.h
@@ -105,12 +105,12 @@ public:
 
   uint64_t cellRowCount() const
   {
-    return gridRowCount() * cell_rows_per_grid;
+    return static_cast<uint64_t>(gridRowCount()) * cell_rows_per_grid;
   }
 
   uint64_t cellColumnCount(uint32_t row) const
   {
-    return gridColumnCount(row) * cell_columns_per_grid;
+    return static_cast<uint64_t>(gridColumnCount(row)) * cell_columns_per_grid;
   }
 
   friend std::ostream& operator<< (std::ostream& stream, const GridBounds& bounds)
